Move ComponentClassName to types
