+++ packages/react/src/components/AccountSettings/ChangePassword/ChangePassword.tsx
@@ -13,8 +13,12 @@ import {
 
 import { useAuth } from '../../../internal';
 import { View, Flex } from '../../../primitives';
-import { ComponentClassName } from '../constants';
-import { FormValues, BlurredFields, ValidationError } from '../types';
+import {
+  ComponentClassName,
+  FormValues,
+  BlurredFields,
+  ValidationError,
+} from '../types';
 import { ChangePasswordProps, ValidateParams } from './types';
 import DEFAULTS from './defaults';
