--- packages/react/src/components/AccountSettings/constants.ts
@@ -1,5 +1,0 @@
-export enum ComponentClassName {
-  ChangePassword = 'amplify-accountsettings-changepassword',
-  DeleteUser = 'amplify-accountsettings-deleteuser',
-  SetupTOTP = 'amplify-accountsettings-setuptotp',
-}
